Add sourcemappath option that is used as relpath for files in sourcemap.
Sass creates filepaths in sourcemap file relative to the output file. If
we use a tempfile as output the generated sourcemap will contain paths
relative to that file.

We add an option `sourcemappath` that, if set, will be the path where
sass creates its outputfile, with the effect that the paths in the
sourcemap will be relative to the given path.