Conditionalize error message trimming, allow node error through noLogMatcher
Trim was blowing up when an error object was passed to it. Sindre might have
noticed that this would happen earlier :sheep:.

Allow the node error message through so users and maintainers have more info on
what's happening, but still display the missing deps warning as this is the
reason for most (all?) 'spawn ENOENT' errors.