Add loadPaths for each intermediate file
0.5.0 processes each file separately, adding a loadPath for it so it acts like
it's being processed in place. This re-implements the feature but for all the
files being processed by update at once.

We get this information by using a version of gulp-intermediate that allows
access to the original vinyl objects. Not sure how holding those in memory is
going to affect performance on low powered devices, but for people processing
100s of kb of Sass that's probably the least of their worries.

Testing and probably refinement is necessary, but let's make this available and
see if this alleviates some of the pain users are feeling.