Update index.js
When I would pass in 'inline' for 'sourcemap' option gulp would create a file named 'inline' which contained an error. This was because the formatting of the argument was incorrect. My proposed fix should work if a valid sourcemap is passed in. The argument is then passed to sass appropriately. Currently a deprecation warning is also logged to STDOUT if one passes in {sourcemap:true,sourcemapMap:''}.