Fix primary test, rearrange fixtures, add results files
Writing output with string concatenation in test.js is messy and doesn't scale.
Puts source fixtures in "source: and adds a "results" dir with expected compiled
CSS files.

Because the erroring file will have different stack traces per environment, we
need to test that it exists in a roundabout way. Could be cleaner.