Remove falsely passing error test
During this test if the error is never emitted the assertion is never checked.
This has been passing falsely since we removed checking for the string
'(E|e)rror' at the beginning of stdout|err, and then throwing that output as an
error.

As a reminder, we removed that functionality because:

1. Compile errors are considered expected output / standard logging in Sass
   https://github.com/sass/sass/issues/1715#issuecomment-100401265
2. We can't guarantee that Sass's 'Error' log will be at the beginning of a log
	 message. Log messages are being split and combined non-deterministically by
	 either Sass itself or node reading the CLI output.
	 https://github.com/sass/sass/issues/1715#issuecomment-98213569

If we read per line we could add individual Sass errors -> gulp errors back in.
Feature for another day.