Move variable creation to place of assignment
I used to prefer declaring all my vars at the top of my function scope, but now
find those extra lines hard to parse and unnecessary with strict mode. Strict
mode protects us from accidentally assigning a global var so we can be sure that
all of these assignments are the first instance of their use in the file.