Remove recompiling instructions from readme, add error to task
Because we have a good amount of users moving from design to development we've
been overly explicit in instructions. But I'd also like to reduce the cognitive
load for everyone by keeping the readme short and to the point. gulp users
should know how to use gulp to watch their files, and if not now is a good time
to learn.

Specifically, removed the recompiling instructions from the readme and added the
information as an error thrown if users try to set watching flags for the
plugin.