Remove redundant Sass docs
Sass options change over time. IMO it's not a good use of the maintainers time
to keep this list up to date, and the increased readme size deemphasizes
important plugin documentation. Instead let's provide usage examples and teach
our users to fish.

Note that gulp-sass takes a similar approach. If support requests grow as a
result of the removal we can add them back.