Add "tempDir" test
Right now we need to set this test up a little differently: setup and assertion
in the it block, because the tempDir is removed when the sass function is done.
When we implement caching (currently active PR) the tempDir will not be cleared
and we can refactor this test to split setup and assertions.