Add concurrency test
Ensure that our automatic unique intermediate directories per task are working.
If files were all compiled to the same directory it is likely that the results
of one task would be streamed through with the results of another (as has
happened many times in the past when users don't know about the container
option; see closed issues).

Note that it's not *guaranteed* that this would happen... but the possibility of
catching something through an intermittent test failure is good enough for me
right now.