Add glob source test
These tests and the directory tests work but could be better. The forEach tests
pass when no files are pushed through the stream, and rely on the "number of
files" assertion to identify a no-files-in-the-stream error. Really all of the
tests should fail if no files are pushed into the stream.

These should be refactored in the near future.