Add sourcemap tests for glob source
Sourcemaps are rewritten using the base directory variable, which is calculated
differently depending on whether the source is a file, directory, or glob. I
want to make sure this is working, and if an edge case is discovered we have a
scaffold ready to go to write a new test against.