Implement array sources, remove directory sources
This makes everything work more like gulp.src. Definitely breaking changes. The
logic for maintaining support for directory sources would be complex, and I
don't think it's worth it; working like gulp.src is the primary goal.

This commit is a little chunky. OH WELL, it is my free time and I've been so
good otherwise.